+++ R/put_object.R
@@ -1,23 +1,33 @@
+#' @rdname put_object
 #' @title Put object
 #' @description Puts an object into an S3 bucket
 #' @param file A character string containing the filename (or full path) of the file you want to upload to S3. Alternatively, an raw vector containing the file can be passed directly, in which case \code{object} needs to be specified explicitly.
-#' @template bucket
 #' @param object A character string containing the name the object should have in S3 (i.e., its "object key"). If missing, the filename is used.
-#' @template dots
+#' @param folder A character string containing a folder name. (A trailing slash is not required.)
+#' @template bucket
 #' @param multipart A logical indicating whether to use multipart uploads. See \url{http://docs.aws.amazon.com/AmazonS3/latest/dev/mpuoverview.html}. If \code{file} is less than 100 MB, this is ignored.
-#' @param headers List of request headers for the REST call.   
-#' @details This provide a generic interface for sending files (or serialized, in-memory representations thereof) to S3. Some convenience wrappers are provided for common tasks: \code{\link{s3save}} and \code{\link{s3saveRDS}}.
+#' @template acl
+#' @param headers List of request headers for the REST call.
+#' @template dots
+#' @details This provide a generic interface for sending files (or serialized, in-memory representations thereof) to S3. Some convenience wrappers are provided for common tasks: e.g., \code{\link{s3save}} and \code{\link{s3saveRDS}}.
+#' 
+#' Note that S3 is a flat file store. So there is no folder hierarchy as in a traditional hard drive. However, S3 allows users to create pseudo-folders by prepending object keys with \code{foldername/}. The \code{put_folder} function is provided as a high-level convenience function for creating folders. This is not actually necessary as objects with slashes in their key will be displayed in the S3 web console as if they were in folders, but it may be useful for creating an empty directory (which is possible in the web console).
 #' 
 #' @return If successful, \code{TRUE}.
 #' @examples
 #' \dontrun{
 #'   library("datasets")
-#' 
+#'   
 #'   # write file to S3
 #'   tmp <- tempfile()
 #'   on.exit(unlink(tmp))
 #'   utils::write.csv(mtcars, file = tmp)
 #'   put_object(tmp, object = "mtcars.csv", bucket = "myexamplebucket")
+#' 
+#'   # create a "folder" in a bucket
+#'   put_folder("example", bucket = "myexamplebucket")
+#'   ## write object to the "folder"
+#'   put_object(tmp, object = "example/mtcars.csv", bucket = "myexamplebucket")
 #' 
 #'   # write serialized, in-memory object to S3
 #'   x <- rawConnection(raw(0), "w")
@@ -40,7 +50,16 @@
 #' @seealso \code{\link{put_bucket}}, \code{\link{get_object}}, \code{\link{delete_object}}
 #' @importFrom utils head
 #' @export
-put_object <- function(file, object, bucket, multipart = FALSE, headers = list(), ...) {
+put_object <- 
+function(file, 
+         object, 
+         bucket, 
+         multipart = FALSE, 
+         acl = c("private", "public-read", "public-read-write", 
+                 "aws-exec-read", "authenticated-read", 
+                 "bucket-owner-read", "bucket-owner-full-control"),
+         headers = list(), 
+         ...) {
     if (missing(object) && is.character(file)) {
         object <- basename(file)
     } else {
@@ -49,6 +68,8 @@ put_object <- function(file, object, bucket, multipart = FALSE, headers = list()
         }
         object <- get_objectkey(object)
     }
+    acl <- match.arg(acl)
+    headers <- c(list(`x-amz-acl` = acl), headers)
     if (isTRUE(multipart)) {
         if (is.character(file) && file.exists(file)) {
             file <- readBin(file, what = "raw")
@@ -77,7 +98,7 @@ put_object <- function(file, object, bucket, multipart = FALSE, headers = list()
         }
         
         # initialize the upload
-        initialize <- post_object(file = NULL, object = object, bucket = bucket, query = list(uploads = ""), ...)
+        initialize <- post_object(file = NULL, object = object, bucket = bucket, query = list(uploads = ""), headers = headers, ...)
         id <- initialize[["UploadId"]]
         
         # loop over parts
@@ -86,7 +107,7 @@ put_object <- function(file, object, bucket, multipart = FALSE, headers = list()
         for (i in seq_along(parts)) {
             query <- list(partNumber = i, uploadId = id)
             r <- try(put_object(file = parts[[i]], object = object, bucket = bucket, 
-                                multipart = FALSE, query = query), 
+                                multipart = FALSE, headers = headers, query = query), 
                      silent = FALSE)
             if (inherits(r, "try-error")) {
                 abort(id)
@@ -112,6 +133,15 @@ put_object <- function(file, object, bucket, multipart = FALSE, headers = list()
                     ...)
         return(TRUE)
     }
+}
+
+#' @rdname put_object
+#' @export
+put_folder <- function(folder, bucket, ...) {
+    if (!endsWith(folder, "/")) {
+        folder <- paste0(folder, "/")
+    }
+    put_object(raw(0), object = folder, bucket = bucket, ...)
 }
 
 post_object <- function(file, object, bucket, headers = list(), ...) {