+++ NEWS.md
@@ -1,3 +1,18 @@
+# CHANGES to aws.s3 0.3.3
+
+* `put_object()` and `put_bucket() now expose explicit `acl` arguments. (#137)
+* `get_acl()` and `put_acl()` are now exported. (#137)
+* Added a high-level `put_folder()` convenience function for creating an empty pseudo-folder.
+
+# CHANGES to aws.s3 0.3.2
+
+* `put_bucket()` now errors if the request is unsuccessful. (#132, h/t Sean Kross)
+* Fixed a bug in the internal function `setup_s3_url()` when `region = ""`.
+
+# CHANGES to aws.s3 0.3.1
+
+* DESCRIPTION file fix for CRAN.
+
 # CHANGES to aws.s3 0.3.0
 
 * CRAN (beta) release. (#126)